backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
drf-extensions